Remove dependency between detekt-tooling and detekt-api
