+++ detekt-common-api/config/detekt.yml
@@ -1,0 +1,3 @@
+libraries:
+  ForbiddenPublicDataClass:
+    active: true
