+++ detekt-common-api/src/main/kotlin/dev/detekt/api/Detektion.kt
