+++ detekt-common-api/src/main/kotlin/dev/detekt/api/Notification.kt
