+++ detekt-common-api/src/main/kotlin/dev/detekt/api/ProjectMetric.kt
